Refactor code and improve GUI
