--- src/main/java/Exceptions/DukeException.java
@@ -1,10 +1,0 @@
-package Exceptions;
-
-/**
- * DukeException is a class that deals with exceptions specific to the DukeBot
- */
-public class DukeException extends Exception {
-    public DukeException(String message) {
-        super(message);
-    }
-}
