--- src/main/java/Main.java
@@ -1,10 +1,0 @@
-import javafx.application.Application;
-
-/**
- * A launcher class to workaround classpath issues.
- */
-public class Main {
-    public static void main(String[] args) {
-        Application.launch(DukeGui.class, args);
-    }
-}
