+++ src/main/java/duke/Exceptions/DukeException.java
@@ -1,0 +1,10 @@
+package duke.Exceptions;
+
+/**
+ * DukeException is a class that deals with exceptions specific to the DukeBot
+ */
+public class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
