+++ src/main/java/duke/Gui/DukeDialogBox.java
@@ -1,0 +1,50 @@
+package duke.Gui;
+
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.geometry.Pos;
+import javafx.scene.Node;
+import javafx.scene.control.Label;
+import javafx.scene.image.ImageView;
+import javafx.scene.layout.HBox;
+
+/**
+ * DukeDialogBox is a class that formats the Duke and User dialog box
+ */
+public class DukeDialogBox extends HBox {
+
+    private Label text;
+    private ImageView displayPicture;
+
+    public DukeDialogBox(Label l, ImageView iv) {
+        this.text = l;
+        this.displayPicture = iv;
+
+        text.setWrapText(true);
+        displayPicture.setFitWidth(100.0);
+        displayPicture.setFitHeight(100.0);
+
+        this.setAlignment(Pos.TOP_RIGHT);
+        boolean b = this.getChildren().addAll(text, displayPicture);
+        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
+        tmp.get(0).setStyle("-fx-background-color: #FFFFFF; -fx-background-radius: 10; -fx-padding: 10;");
+    }
+
+    private void flip() {
+        this.setAlignment(Pos.TOP_LEFT);
+        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
+        tmp.get(0).setStyle("-fx-background-color: #FFFFFF; -fx-background-radius: 10; -fx-padding: 10;");
+        FXCollections.reverse(tmp);
+        this.getChildren().setAll(tmp);
+    }
+
+    public static DukeDialogBox getUserDialog(Label l, ImageView iv) {
+        return new DukeDialogBox(l, iv);
+    }
+
+    public static DukeDialogBox getDukeDialog(Label l, ImageView iv) {
+        var db = new DukeDialogBox(l, iv);
+        db.flip();
+        return db;
+    }
+}
